+++ app/src/main/java/fr/free/nrw/commons/auth/LoginActivity.java
@@ -411,7 +411,6 @@ public class LoginActivity extends AccountAuthenticatorActivity {
     }
 
     public void startMainActivity() {
-        // TODO move to base activity
         ActivityUtils.startActivityWithFlags(this, MainActivity.class, Intent.FLAG_ACTIVITY_SINGLE_TOP);
         finish();
     }
