+++ app/src/main/java/fr/free/nrw/commons/nearby/NearbyNotificationCardView.java
@@ -14,12 +14,11 @@ import androidx.annotation.Nullable;
 
 import fr.free.nrw.commons.R;
 import fr.free.nrw.commons.contributions.MainActivity;
+import fr.free.nrw.commons.navtab.NavTab;
 import fr.free.nrw.commons.nearby.fragments.NearbyParentFragment;
 import fr.free.nrw.commons.utils.SwipableCardView;
 import fr.free.nrw.commons.utils.ViewUtil;
 import timber.log.Timber;
-
-import static fr.free.nrw.commons.contributions.MainActivity.NEARBY_TAB_POSITION;
 
 /**
  * Custom card view for nearby notification card view on main screen, above contributions list
@@ -86,13 +85,14 @@ public class NearbyNotificationCardView extends SwipableCardView {
 
     private void setActionListeners(Place place) {
         this.setOnClickListener(view -> {
-            MainActivity m = (MainActivity) getContext();
+            MainActivity context = (MainActivity) getContext();
 
             // Change to nearby tab
             //m.viewPager.setCurrentItem(NEARBY_TAB_POSITION);
 
             // Center the map to the place
             //TODO ((NearbyParentFragment) m.contributionsActivityPagerAdapter.getItem(NEARBY_TAB_POSITION)).centerMapToPlace(place);
+            context.centerMapToPlace(place);
         });
     }
